Move nested SingleBiomeProvider to super class
